feat(report): 결과지 생성시 결과 정보 추가, 결과지 생성 코드 변경, 난소암 누락 추가

+++ report/src/main/kotlin/com/greencross/lims/service/report/ReportRouter.kt
@@ -45,7 +45,7 @@ class ReportRouter(
         val service = request.pathVariable("service")
         val lang    = request.pathVariable("lang")
         val batch   = request.pathVariable("batch")
-        val row     = request.pathVariable("row")
+        val row     = request.pathVariable("row").toLong()
         return request.bodyToMono(String::class.java)
             .switchIfEmpty(Mono.just("최초보고"))
             .flatMap { handler.print(sample.toLong(), service, batch, row, lang, it) }
